move app.main.dart into lib folder
