--- at_app_flutter/lib/src/templates/app/main.dart
@@ -1,109 +1,0 @@
-import 'dart:async';
-
-import 'package:flutter/material.dart';
-import 'package:at_client_mobile/at_client_mobile.dart';
-import 'package:at_onboarding_flutter/at_onboarding_flutter.dart'
-    show Onboarding;
-import 'package:at_utils/at_logger.dart' show AtSignLogger;
-import 'package:path_provider/path_provider.dart'
-    show getApplicationSupportDirectory;
-import 'package:at_app_flutter/at_app_flutter.dart' show AtEnv;
-
-Future<void> main() async {
-  await AtEnv.load();
-  runApp(const MyApp());
-}
-
-Future<AtClientPreference> loadAtClientPreference() async {
-  var dir = await getApplicationSupportDirectory();
-  return AtClientPreference()
-        ..rootDomain = AtEnv.rootDomain
-        ..namespace = AtEnv.appNamespace
-        ..hiveStoragePath = dir.path
-        ..commitLogPath = dir.path
-        ..isLocalStoreRequired = true
-      // TODO set the rest of your AtClientPreference here
-      ;
-}
-
-class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
-  @override
-  _MyAppState createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  // * load the AtClientPreference in the background
-  Future<AtClientPreference> futurePreference = loadAtClientPreference();
-  AtClientPreference? atClientPreference;
-
-  final AtSignLogger _logger = AtSignLogger(AtEnv.appNamespace);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      // * The onboarding screen (first screen)
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('MyApp'),
-        ),
-        body: Builder(
-          builder: (context) => Center(
-            child: ElevatedButton(
-              onPressed: () async {
-                var preference = await futurePreference;
-                setState(() {
-                  atClientPreference = preference;
-                });
-                Onboarding(
-                  context: context,
-                  atClientPreference: atClientPreference!,
-                  domain: AtEnv.rootDomain,
-                  rootEnvironment: AtEnv.rootEnvironment,
-                  appAPIKey: AtEnv.appApiKey,
-                  onboard: (value, atsign) {
-                    _logger.finer('Successfully onboarded $atsign');
-                  },
-                  onError: (error) {
-                    _logger.severe('Onboarding throws $error error');
-                  },
-                  nextScreen: const HomeScreen(),
-                );
-              },
-              child: const Text('Onboard an @sign'),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-//* The next screen after onboarding (second screen)
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    /// Get the AtClientManager instance
-    var atClientManager = AtClientManager.getInstance();
-
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Dashboard'),
-      ),
-      body: Center(
-        child: Column(
-          children: [
-            const Text(
-                'Successfully onboarded and navigated to FirstAppScreen'),
-
-            /// Use the AtClientManager instance to get the current atsign
-            Text(
-                'Current @sign: ${atClientManager.atClient.getCurrentAtSign()}'),
-          ],
-        ),
-      ),
-    );
-  }
-}
